requirements.txt
streamlit==1.36.0
pandas==2.2.3
googlemaps==4.10.0
gspread==6.1.4
google-auth==2.36.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
requests==2.32.3 